requirements.txt
streamlit==1.45.1
PyPDF2==3.0.1 
langchain==0.3.25 
langchain-openai==0.3.17
langchain-community==0.3.24 
python-dotenv==1.1.0